Fix compile error when enabling C++11
/home/krf/devel/src/icemon/src/views/starview.h:77:37: error:
‘constexpr’ needed for in-class initialization of static data member
‘const qreal HostItem::HaloMargin’ of non-integral type [-fpermissive]

+++ src/views/starview.h
@@ -74,7 +74,7 @@ private:
 class HostItem : public QGraphicsItemGroup
 {
 public:
-    static const qreal HaloMargin = 4;
+    static const int HaloMargin = 4;
     static const int PenDarkerFactor = 125;
 
     enum { RttiHostItem = 1000 };
